Refactor : .png에서 .svg로 이미지 확장자 변경
