+++ src/components/AlbumIcon/AlbumIcon.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import iconPostAlbumOff from 'assets/img/icon-post-album-off.png';
-import iconPostAlbumOn from 'assets/img/icon-post-album-on.png';
+import iconPostAlbumOff from 'assets/img/icon-post-album-off.svg';
+import iconPostAlbumOn from 'assets/img/icon-post-album-on.svg';
 import BtnStyle from './AlbumIconStyle';
 
 export default function AlbumIcon({ toggle, onclick }) {
