+++ src/components/Comment/Comment.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import deleteComment from 'api/Comment/deleteComment';
 import reportComment from 'api/Comment/reportComment';
-import iconMoreImg from 'assets/img/icon-more.png';
+import iconMoreImg from 'assets/img/icon-more.svg';
 import CommentModal from '../common/Modal/CommentModel/CommentModal';
 import ProfileImg from '../common/ProfileImg/ProfileImg';
 import {
