+++ src/components/CommentInput/CommentInput.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
-import BasicProfile from 'assets/img/basic-profile.png';
+import BasicProfile from 'assets/img/basic-profile.svg';
 import postComment from 'api/Comment/postComment';
 import ProfileImg from '../common/ProfileImg/ProfileImg';
 import {
