+++ src/components/ListIcon/ListIcon.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import IconPostListOn from 'assets/img/icon-post-list-on.png';
-import iconPostListOff from 'assets/img/icon-post-list-off.png';
+import IconPostListOn from 'assets/img/icon-post-list-on.svg';
+import iconPostListOff from 'assets/img/icon-post-list-off.svg';
 import BtnStyle from './ListIconStyle';
 
 export default function ListIcon({ toggle, onclick }) {
