+++ src/components/PostCard/PostCardStyle.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import IconMore from '../../assets/img/icon-more.png';
+import IconMore from 'assets/img/icon-more.svg';
 
 export const PostAccountLiStyle = styled.li``;
