+++ src/components/common/Header/Header.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import iconArrowLeft from 'assets/img/icon-arrow-left.png';
-import iconMore from 'assets/img/icon-more.png';
-import iconSearch from 'assets/img/icon-search.png';
+import iconArrowLeft from 'assets/img/icon-arrow-left.svg';
+import iconMore from 'assets/img/icon-more.svg';
+import iconSearch from 'assets/img/icon-search.svg';
 import Button from '../Button/Button';
 import LoginModal from '../Modal/LoginModalAlert/LoginModal';
 import {
