+++ src/components/common/Nav/Nav.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import styled from 'styled-components';
-import iconHome from 'assets/img/icon-home.png';
-import iconHomeFill from 'assets/img/icon-home-fill.png';
-import iconMessage from 'assets/img/icon-message-circle.png';
-import iconMessageFill from 'assets/img/icon-message-circle-fill.png';
-import iconEdit from 'assets/img/icon-edit.png';
-import iconUser from 'assets/img/icon-user.png';
-import iconUserFill from 'assets/img/icon-user-fill.png';
+import iconHome from 'assets/img/icon-home.svg';
+import iconHomeFill from 'assets/img/icon-home-fill.svg';
+import iconMessage from 'assets/img/icon-message-circle.svg';
+import iconMessageFill from 'assets/img/icon-message-circle-fill.svg';
+import iconEdit from 'assets/img/icon-edit.svg';
+import iconUser from 'assets/img/icon-user.svg';
+import iconUserFill from 'assets/img/icon-user-fill.svg';
 import { ContainerNav, NavUl } from './NavStyle';
 
 export const NavLinkStyle = styled(NavLink)`
