+++ src/pages/NotFound/NotFound.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import icon404Img from 'assets/img/icon-404.png';
+import icon404Img from 'assets/img/icon-404.svg';
 import Blank from 'components/Blank/Blank';
 import { ContainerSectionStyle, H2IR } from './NotFoundStyle';
